+++ models/mysql/couponAssign.js
@@ -37,7 +37,7 @@ const CouponAssign = db.defineModel('couponAssigns', {
         defaultValue: ''
     },
     reducePrice: {
-        type: db.DECIMAL(10, 2),
+        type: db.FLOAT,
         defaultValue: 0
     },
     couponStartDate: {
