+++ models/mysql/trainClass.js
@@ -53,11 +53,11 @@ const TrainClass = db.defineModel('trainClasss', {
         defaultValue: 0
     }, //共多少课时
     trainPrice: {
-        type: db.DECIMAL(10, 2),
+        type: db.FLOAT,
         defaultValue: 0
     },
     materialPrice: {
-        type: db.DECIMAL(10, 2),
+        type: db.FLOAT,
         defaultValue: 0
     },
     teacherId: {
